refactoring structure of user app. Tested. All working
