+++ grabber/grabber/settings.py
@@ -82,7 +82,7 @@ INSTALLED_APPS = [
 
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': [
-        'users.authentication.JWTAuthFromCookie',
+        'users.authentication.authentication.JWTAuthFromCookie',
         'rest_framework_simplejwt.authentication.JWTAuthentication',
 
     ],
@@ -189,6 +189,5 @@ DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 AUTH_USER_MODEL = 'users.CustomUser'
 
 AUTHENTICATION_BACKENDS = [
-    'users.backends.EmailBackend',
     'django.contrib.auth.backends.ModelBackend',
 ]
