--- grabber/users/backends.py
@@ -1,12 +1,0 @@
-from django.contrib.auth.backends import ModelBackend
-from users.models import CustomUser
-
-class EmailBackend(ModelBackend):
-    def authenticate(self, request, email=None, password=None, **kwargs):
-        email = kwargs.get('email', email)
-        try:
-            user = CustomUser.objects.get(email=email)
-            if user.check_password(password):
-                return user
-        except CustomUser.DoesNotExist:
-            return None
