+++ grabber/users/urls.py
@@ -1,27 +1,27 @@
-from django.urls import path
+
 from django.urls import path
 from .views import (
-    AsyncCookieViewLogin,
-    AsyncCookieViewRegister,
-    AsyncCookieViewRefresh,
-    AsyncCookieViewLogout,
-    UserProfileView,
-    AsyncForgotPasswordView,
-    AsyncResetPasswordView
+    login,
+    register,
+    refresh,
+    logout,
+    user_profile,
+    forgot_password,
+    reset_password
 )
 
 app_name = 'users'  # дозволяє використовувати простір імен при реверсі
 
 urlpatterns = [
     # 🔐 Auth endpoints
-    path('login/', AsyncCookieViewLogin.as_view(), name='login'),
-    path('register/', AsyncCookieViewRegister.as_view(), name='register'),
-    path('refresh/', AsyncCookieViewRefresh.as_view(), name='refresh'),
-    path('logout/', AsyncCookieViewLogout.as_view(), name='logout'),
+    path('login/', login.AsyncCookieViewLogin.as_view(), name='login'),
+    path('register/', register.AsyncCookieViewRegister.as_view(), name='register'),
+    path('refresh/', refresh.AsyncCookieViewRefresh.as_view(), name='refresh'),
+    path('logout/', logout.AsyncCookieViewLogout.as_view(), name='logout'),
 
     # 👤 User profile
-    path('profile/', UserProfileView.as_view(), name='profile'),
+    path('profile/', user_profile.AsyncUserProfileView.as_view(), name='profile'),
 
-    path('forgot-password/', AsyncForgotPasswordView.as_view(), name='forgot-password'),
-    path('reset-password/', AsyncResetPasswordView.as_view(), name='reset-password'),
+    path('forgot-password/', forgot_password.AsyncForgotPasswordView.as_view(), name='forgot-password'),
+    path('reset-password/', reset_password.AsyncResetPasswordView.as_view(), name='reset-password'),
 ]
